requirements.txt
numpy
pandas
seaborn
ydata_profiling
mrmr_selection
scikit-learn
xgboost
catboost
# -e .